@@ -12,10 +12,16 @@ class FlashChat extends StatelessWidget {
     return MaterialApp(
       theme: ThemeData.dark().copyWith(
         textTheme: TextTheme(
-          body1: TextStyle(color: Colors.black54),
+          bodyLarge: TextStyle(color: Colors.black54),
         ),
       ),
-      home: WelcomeScreen(),
+      initialRoute: WelcomeScreen.ID,
+      routes: {
+        WelcomeScreen.ID: (context) => WelcomeScreen(),
+        LoginScreen.ID: (context) => LoginScreen(),
+        RegistrationScreen.ID: (context) => RegistrationScreen(),
+        ChatScreen.ID: (context) => ChatScreen(),
+      },
     );
   }
 }
